Define the methods to type case to a hash
This will be used to save the file. As the file needs to be safe loaded,
Pathname objects must not appear in the hash. They are removed during
the type conversion

Also always default to the empty string path if the values are missing.
This ensure their are always `Pathname`